+++ Teleopti.Ccc.Web/Teleopti.Ccc.Web/Areas/MyTime/Core/WeekSchedule/Mapping/WeekScheduleDayDomainData.cs
@@ -13,6 +13,6 @@ namespace Teleopti.Ccc.Web.Areas.MyTime.Core.WeekSchedule.Mapping
 		public IVisualLayerCollection Projection { get; set; }
         public IVisualLayerCollection ProjectionYesterday { get; set; }
         public TimePeriod MinMaxTime { get; set; }
-		public int Allowance { get; set; }
+		public double Allowance { get; set; }
 	}
 }
